\name{Leukemia}
\alias{Leukemia}
\docType{data}
\title{
Data from Golub et al. 1999
}
\description{
Data from Golub, Todd R., et al. "Molecular classification of cancer: 
class discovery and class prediction by gene expression monitoring." 
Science 286.5439 (1999): 531-537.}

\usage{data(Leukemia)}

\keyword{datasets}
